changed the way unit tests are run so they can have a default adapter
Previously, the code that implemented the test_* targets assumed that
tests would always be run using one of those targets so it didn't have
a default adapter.  This didn't work well for people that like to run
a test at a time using bare Ruby.  This commit modifies the test
targets so that they set the ADAPTER environment variable and then run
the tests.  abstract_unit.rb, which loads the appropriate adapter,
uses this environment variable to decide which adapter to load, and
defaults to postgresql.  If you want to run a test at a time you can
set ADAPTER="adapter" on the command line.

This commit also changes the test target names to make them more
consistent with the other per-database targets, so what used to be
test_mysql is now mysql:test.