Use ADAPTER env variable to select database to test.
Commit 579c29fe0ee9a881ca3ecf89f58820796a716a26 changed the test setup in
abstract_unit so that it used the first adapter in databases.yml
(as presented by the YAML parser; not necessarily the first one in the file),
rather than looking at the ADAPTER env variable. This change broke the
database-specific rake tasks (they would run against the database selected
by databases.yml rather than the one the user requested) and made parts of the
testing README incorrect.

This commit restores the ability to use the ADAPTER env var to select a
database to test against. This works in both the rake tasks and when running
from the command line:

    ADAPTER=sqlite3 ruby test/test_suite.rb

The default adapter is `postgresql`, as was previously the case.

Incidentally, this commit also removes & ignores the committed
test/connections/databases.yml file in preference to the (existing)
databases.examples.yml file. This change will make it easier for contributors
to avoid accidentally contributing changes to the test database configuration.