Allow AR logs to go to a file when running tests.
The test runner configures ActiveRecord to send its messages to STDOUT. This
is helpful for debugging, but a lot of noise if you're just trying to see if
everything passed. This commit allows you to specify a file to use instead
via the CPK_LOGFILE environment variable. (STDOUT is still the default.)