Standardize on integer PKs in polymorphic fixtures.
The mix of string and integer PKs and FKs in the comments table and the three
tables (employees, users, and hacks) to which it refers causes problems in
databases that do strict type checking in SELECTs. (Problems observed on
PostgreSQL 8.3+ and Oracle.) This commit makes all the FKs and PKs integers to
eliminate these problems.

N.b.: this commit includes changes to the test database schema; you will need
to rebuild your test database after applying it. There are no changes for the
db2 schema because it does not include the relevant tables. The changes to
the MySQL schema were not tested.