Fix test for bracket assignment (see description)
Rails 4 has a commit (see https://github.com/rails/rails/commit/522c0fdb555d54ca68d1acfd1bd8e67fd0040086) that ensures the PK is always present in the @attributes hash, which was causing this test to break when attempting to assign values to that key in the hash (since it just loops over them).

Given there was no expectation the key existed in the prior versions, a valid fix seemed to me to be to do not test assignment in this version.

However, I'm not sure if that's a valid assumption or not, so please review / consider that carefully.