Fix habtm association #delete_records
Fixes #273 - habtm association issue when destroying a record.

Let say we have a Foo record that has-and-belongs-to-many Bar records.
When we destroy Foo record, the association data in the junction table
bars_foos doesn't get deleted. On closer inspection this happens because
of the invalid delete statement which silently fails, e.g.:
  DELETE FROM `bars_foos` WHERE `bars_foos`.`foo_id` = 1 AND NULL

This commit amends the habtm test to fail and demonstrate this bug.

Then it addresses this bug inside the HasAndBelongsToManyAssociation in
the #delete_records. It makes sure only non-empty predicates are used
when scoping the to-be-removed records of the junction table.

Consequently a valid delete statement is generated and test passes.