Update persistence.rb
I'm using CPK to make Type II tables that use a start_datetime and end_datetime.  The primary key is a composite of ID and start_datetime.  Because the rails "id" is now mapped to the array of key attributes, and in many cases I want to directly reference the integer ID number, I ditched the normal ID built by a rails migration with "id: false".  I then created an xid column that has a Postgres sequence (in other RDBMS it's AUTO_INCREMENT / identity, etc).  So I can refer to the composite key with "id", or just the integer part with "xid".  When calling .create this was coming back with nil for the xid unless I take off memoization, as you see on line 12.